Recipes Project/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-login
werkzeug
mysql-connector-python 